+++ relay-server/src/utils/rate_limits.rs
@@ -11,7 +11,7 @@ use smallvec::SmallVec;
 
 use crate::envelope::{AttachmentParentType, AttachmentType, Envelope, Item, ItemType};
 use crate::integrations::Integration;
-use crate::managed::{Managed, ManagedEnvelope};
+use crate::managed::Managed;
 use crate::services::outcome::Outcome;
 
 /// Name of the rate limits header.
@@ -667,18 +667,6 @@ impl Enforcement {
     /// 1. The item remains in the envelope.
     /// 2. Enforcements are empty. Rate limiting has occurred at an earlier stage in the pipeline.
     /// 3. Rate limits are empty.
-    pub fn apply_with_outcomes(self, envelope: &mut ManagedEnvelope) {
-        envelope
-            .envelope_mut()
-            .retain_items(|item| self.retain_item(item));
-        self.track_outcomes(envelope);
-    }
-
-    /// Applies the [`Enforcement`] on the [`Envelope`] by removing all items that were rate limited
-    /// and emits outcomes for each rate limited category.
-    ///
-    /// Works exactly like [`Self::apply_with_outcomes`], but instead operates on [`Managed`]
-    /// instead of [`ManagedEnvelope`].
     pub fn apply_to_managed(self, envelope: &mut Managed<Box<Envelope>>) {
         envelope.modify(|envelope, records| {
             envelope.retain_items(|item| self.retain_item(item));
@@ -776,15 +764,6 @@ impl Enforcement {
             | ItemType::Unknown(_) => true,
         }
     }
-
-    /// Invokes track outcome on all enforcements reported by the [`EnvelopeLimiter`].
-    ///
-    /// Relay generally does not emit outcomes for sessions, so those are skipped.
-    fn track_outcomes(self, envelope: &mut ManagedEnvelope) {
-        for (outcome, category, quantity) in self.get_outcomes() {
-            envelope.track_outcome(outcome, category, quantity)
-        }
-    }
 }
 
 /// Which limits to check with the [`EnvelopeLimiter`].
@@ -797,8 +776,6 @@ pub enum CheckLimits {
     /// Additionally even if the item is later dropped by dynamic sampling, it must still be around to extract metrics
     /// and cannot be dropped too early.
     NonIndexed,
-    /// Checks all limits against the envelope.
-    All,
 }
 
 struct Check<F, E, R> {
@@ -861,6 +838,7 @@ where
     /// This ensures that rate limits for the given data category are checked even if there is no
     /// matching item in the envelope. Other items are handled according to the rules as if the
     /// event item were present.
+    #[cfg(test)]
     pub fn assume_event(&mut self, category: DataCategory) {
         self.event_category = Some(category);
     }
@@ -1610,20 +1588,8 @@ mod tests {
                 envelope.add_item(item);
             )*
 
-            let (outcome_aggregator, _) = Addr::custom();
-
-            ManagedEnvelope::new(
-                envelope,
-                outcome_aggregator,
-            )
+            envelope
         }}
-    }
-
-    fn set_extracted(envelope: &mut Envelope, ty: ItemType) {
-        envelope
-            .get_item_by_mut(|item| *item.ty() == ty)
-            .unwrap()
-            .set_metrics_extracted(true);
     }
 
     fn rate_limit(category: DataCategory) -> RateLimit {
@@ -1708,13 +1674,14 @@ mod tests {
 
     async fn enforce_and_apply(
         mock: Arc<Mutex<MockLimiter>>,
-        envelope: &mut ManagedEnvelope,
+        envelope: Box<Envelope>,
         #[allow(unused_variables)] assume_event: Option<DataCategory>,
-    ) -> (Enforcement, RateLimits) {
+    ) -> (Box<Envelope>, Enforcement, RateLimits) {
+        let mut envelope = Managed::from_envelope(envelope, Addr::custom().0);
         let scoping = envelope.scoping();
 
         #[allow(unused_mut)]
-        let mut limiter = EnvelopeLimiter::new(CheckLimits::All, move |s, q| {
+        let mut limiter = EnvelopeLimiter::new(CheckLimits::NonIndexed, move |s, q| {
             let mock = mock.clone();
             async move {
                 let mut mock = mock.lock().await;
@@ -1726,16 +1693,14 @@ mod tests {
             limiter.assume_event(assume_event);
         }
 
-        let (enforcement, limits) = limiter
-            .compute(envelope.envelope_mut(), &scoping)
-            .await
-            .unwrap();
+        let (enforcement, limits) = limiter.compute(&envelope, &scoping).await.unwrap();
 
         // We implemented `clone` only for tests because we don't want to make `apply_with_outcomes`
         // &self because we want move semantics to prevent double tracking.
-        enforcement.clone().apply_with_outcomes(envelope);
+        enforcement.clone().apply_to_managed(&mut envelope);
+        let envelope = envelope.accept(|envelope| envelope);
 
-        (enforcement, limits)
+        (envelope, enforcement, limits)
     }
 
     fn mock_limiter(categories: &[DataCategory]) -> Arc<Mutex<MockLimiter>> {
@@ -1749,61 +1714,61 @@ mod tests {
 
     #[tokio::test]
     async fn test_enforce_pass_empty() {
-        let mut envelope = envelope![];
+        let envelope = envelope![];
 
         let mock = mock_limiter(&[]);
-        let (_, limits) = enforce_and_apply(mock, &mut envelope, None).await;
+        let (envelope, _, limits) = enforce_and_apply(mock, envelope, None).await;
 
         assert!(!limits.is_limited());
-        assert!(envelope.envelope().is_empty());
+        assert!(envelope.is_empty());
     }
 
     #[tokio::test]
     async fn test_enforce_limit_error_event() {
-        let mut envelope = envelope![Event];
+        let envelope = envelope![Event];
 
         let mock = mock_limiter(&[DataCategory::Error]);
-        let (_, limits) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
+        let (envelope, _, limits) = enforce_and_apply(mock.clone(), envelope, None).await;
 
         assert!(limits.is_limited());
-        assert!(envelope.envelope().is_empty());
+        assert!(envelope.is_empty());
         mock.lock().await.assert_call(DataCategory::Error, 1);
     }
 
     #[tokio::test]
     async fn test_enforce_limit_error_with_attachments() {
-        let mut envelope = envelope![Event, Attachment];
+        let envelope = envelope![Event, Attachment];
 
         let mock = mock_limiter(&[DataCategory::Error]);
-        let (_, limits) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
+        let (envelope, _, limits) = enforce_and_apply(mock.clone(), envelope, None).await;
 
         assert!(limits.is_limited());
-        assert!(envelope.envelope().is_empty());
+        assert!(envelope.is_empty());
         mock.lock().await.assert_call(DataCategory::Error, 1);
     }
 
     #[tokio::test]
     async fn test_enforce_limit_minidump() {
-        let mut envelope = envelope![Attachment::Minidump];
+        let envelope = envelope![Attachment::Minidump];
 
         let mock = mock_limiter(&[DataCategory::Error]);
-        let (_, limits) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
+        let (envelope, _, limits) = enforce_and_apply(mock.clone(), envelope, None).await;
 
         assert!(limits.is_limited());
-        assert!(envelope.envelope().is_empty());
+        assert!(envelope.is_empty());
         mock.lock().await.assert_call(DataCategory::Error, 1);
     }
 
     #[tokio::test]
     async fn test_enforce_limit_attachments() {
-        let mut envelope = envelope![Attachment::Minidump, Attachment];
+        let envelope = envelope![Attachment::Minidump, Attachment];
 
         let mock = mock_limiter(&[DataCategory::Attachment]);
-        let (_, limits) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
+        let (envelope, _, limits) = enforce_and_apply(mock.clone(), envelope, None).await;
 
         // Attachments would be limited, but crash reports create events and are thus allowed.
         assert!(limits.is_limited());
-        assert_eq!(envelope.envelope().len(), 1);
+        assert_eq!(envelope.len(), 1);
         mock.lock().await.assert_call(DataCategory::Error, 1);
         mock.lock().await.assert_call(DataCategory::Attachment, 20);
     }
@@ -1811,13 +1776,13 @@ mod tests {
     /// Limit stand-alone profiles.
     #[tokio::test]
     async fn test_enforce_limit_profiles() {
-        let mut envelope = envelope![Profile, Profile];
+        let envelope = envelope![Profile, Profile];
 
         let mock = mock_limiter(&[DataCategory::Profile]);
-        let (enforcement, limits) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
+        let (envelope, enforcement, limits) = enforce_and_apply(mock.clone(), envelope, None).await;
 
         assert!(limits.is_limited());
-        assert_eq!(envelope.envelope().len(), 0);
+        assert_eq!(envelope.len(), 0);
         mock.lock().await.assert_call(DataCategory::Profile, 2);
 
         assert_eq!(
@@ -1834,19 +1799,19 @@ mod tests {
     async fn test_enforce_limit_profile_chunks_no_profile_type() {
         // In this test we have profile chunks which have not yet been classified, which means they
         // should not be rate limited.
-        let mut envelope = envelope![ProfileChunk, ProfileChunk];
+        let envelope = envelope![ProfileChunk, ProfileChunk];
 
         let mock = mock_limiter(&[DataCategory::ProfileChunk]);
-        let (enforcement, limits) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
+        let (envelope, enforcement, limits) = enforce_and_apply(mock.clone(), envelope, None).await;
         assert!(!limits.is_limited());
         assert_eq!(get_outcomes(enforcement), vec![]);
 
         let mock = mock_limiter(&[DataCategory::ProfileChunkUi]);
-        let (enforcement, limits) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
+        let (envelope, enforcement, limits) = enforce_and_apply(mock.clone(), envelope, None).await;
         assert!(!limits.is_limited());
         assert_eq!(get_outcomes(enforcement), vec![]);
 
-        assert_eq!(envelope.envelope().len(), 2);
+        assert_eq!(envelope.len(), 2);
     }
 
     #[tokio::test]
@@ -1855,16 +1820,16 @@ mod tests {
 
         let mut item = Item::new(ItemType::ProfileChunk);
         item.set_platform("python".to_owned());
-        envelope.envelope_mut().add_item(item);
+        envelope.add_item(item);
         let mut item = Item::new(ItemType::ProfileChunk);
         item.set_platform("javascript".to_owned());
-        envelope.envelope_mut().add_item(item);
+        envelope.add_item(item);
 
         let mock = mock_limiter(&[DataCategory::ProfileChunkUi]);
-        let (enforcement, limits) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
+        let (envelope, enforcement, limits) = enforce_and_apply(mock.clone(), envelope, None).await;
 
         assert!(limits.is_limited());
-        assert_eq!(envelope.envelope().len(), 1);
+        assert_eq!(envelope.len(), 1);
         mock.lock()
             .await
             .assert_call(DataCategory::ProfileChunkUi, 1);
@@ -1882,16 +1847,16 @@ mod tests {
 
         let mut item = Item::new(ItemType::ProfileChunk);
         item.set_platform("python".to_owned());
-        envelope.envelope_mut().add_item(item);
+        envelope.add_item(item);
         let mut item = Item::new(ItemType::ProfileChunk);
         item.set_platform("javascript".to_owned());
-        envelope.envelope_mut().add_item(item);
+        envelope.add_item(item);
 
         let mock = mock_limiter(&[DataCategory::ProfileChunk]);
-        let (enforcement, limits) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
+        let (envelope, enforcement, limits) = enforce_and_apply(mock.clone(), envelope, None).await;
 
         assert!(limits.is_limited());
-        assert_eq!(envelope.envelope().len(), 1);
+        assert_eq!(envelope.len(), 1);
         mock.lock()
             .await
             .assert_call(DataCategory::ProfileChunkUi, 1);
@@ -1906,13 +1871,13 @@ mod tests {
     /// Limit replays.
     #[tokio::test]
     async fn test_enforce_limit_replays() {
-        let mut envelope = envelope![ReplayEvent, ReplayRecording, ReplayVideo];
+        let envelope = envelope![ReplayEvent, ReplayRecording, ReplayVideo];
 
         let mock = mock_limiter(&[DataCategory::Replay]);
-        let (enforcement, limits) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
+        let (envelope, enforcement, limits) = enforce_and_apply(mock.clone(), envelope, None).await;
 
         assert!(limits.is_limited());
-        assert_eq!(envelope.envelope().len(), 0);
+        assert_eq!(envelope.len(), 0);
         mock.lock().await.assert_call(DataCategory::Replay, 3);
 
         assert_eq!(get_outcomes(enforcement), vec![(DataCategory::Replay, 3),]);
@@ -1921,13 +1886,13 @@ mod tests {
     /// Limit monitor checkins.
     #[tokio::test]
     async fn test_enforce_limit_monitor_checkins() {
-        let mut envelope = envelope![CheckIn];
+        let envelope = envelope![CheckIn];
 
         let mock = mock_limiter(&[DataCategory::Monitor]);
-        let (enforcement, limits) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
+        let (envelope, enforcement, limits) = enforce_and_apply(mock.clone(), envelope, None).await;
 
         assert!(limits.is_limited());
-        assert_eq!(envelope.envelope().len(), 0);
+        assert_eq!(envelope.len(), 0);
         mock.lock().await.assert_call(DataCategory::Monitor, 1);
 
         assert_eq!(get_outcomes(enforcement), vec![(DataCategory::Monitor, 1)])
@@ -1935,14 +1900,14 @@ mod tests {
 
     #[tokio::test]
     async fn test_enforce_pass_minidump() {
-        let mut envelope = envelope![Attachment::Minidump];
+        let envelope = envelope![Attachment::Minidump];
 
         let mock = mock_limiter(&[DataCategory::Attachment]);
-        let (_, limits) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
+        let (envelope, _, limits) = enforce_and_apply(mock.clone(), envelope, None).await;
 
         // If only crash report attachments are present, we don't emit a rate limit.
         assert!(!limits.is_limited());
-        assert_eq!(envelope.envelope().len(), 1);
+        assert_eq!(envelope.len(), 1);
         mock.lock().await.assert_call(DataCategory::Error, 1);
         mock.lock().await.assert_call(DataCategory::Attachment, 10);
     }
@@ -1954,38 +1919,38 @@ mod tests {
         let mut item = Item::new(ItemType::Attachment);
         item.set_payload(ContentType::OctetStream, "0123456789");
         item.set_rate_limited(true);
-        envelope.envelope_mut().add_item(item);
+        envelope.add_item(item);
 
         let mock = mock_limiter(&[DataCategory::Error]);
-        let (_, limits) = enforce_and_apply(mock, &mut envelope, None).await;
+        let (envelope, _, limits) = enforce_and_apply(mock, envelope, None).await;
 
         assert!(!limits.is_limited()); // No new rate limits applied.
-        assert_eq!(envelope.envelope().len(), 1); // The item was retained
+        assert_eq!(envelope.len(), 1); // The item was retained
     }
 
     #[tokio::test]
     async fn test_enforce_pass_sessions() {
-        let mut envelope = envelope![Session, Session, Session];
+        let envelope = envelope![Session, Session, Session];
 
         let mock = mock_limiter(&[DataCategory::Error]);
-        let (_, limits) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
+        let (envelope, _, limits) = enforce_and_apply(mock.clone(), envelope, None).await;
 
         // If only crash report attachments are present, we don't emit a rate limit.
         assert!(!limits.is_limited());
-        assert_eq!(envelope.envelope().len(), 3);
+        assert_eq!(envelope.len(), 3);
         mock.lock().await.assert_call(DataCategory::Session, 3);
     }
 
     #[tokio::test]
     async fn test_enforce_limit_sessions() {
-        let mut envelope = envelope![Session, Session, Event];
+        let envelope = envelope![Session, Session, Event];
 
         let mock = mock_limiter(&[DataCategory::Session]);
-        let (_, limits) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
+        let (envelope, _, limits) = enforce_and_apply(mock.clone(), envelope, None).await;
 
         // If only crash report attachments are present, we don't emit a rate limit.
         assert!(limits.is_limited());
-        assert_eq!(envelope.envelope().len(), 1);
+        assert_eq!(envelope.len(), 1);
         mock.lock().await.assert_call(DataCategory::Error, 1);
         mock.lock().await.assert_call(DataCategory::Session, 2);
     }
@@ -1993,37 +1958,37 @@ mod tests {
     #[tokio::test]
     #[cfg(feature = "processing")]
     async fn test_enforce_limit_assumed_event() {
-        let mut envelope = envelope![];
+        let envelope = envelope![Transaction];
 
         let mock = mock_limiter(&[DataCategory::Transaction]);
-        let (_, limits) =
-            enforce_and_apply(mock.clone(), &mut envelope, Some(DataCategory::Transaction)).await;
+        let (envelope, _, limits) =
+            enforce_and_apply(mock.clone(), envelope, Some(DataCategory::Transaction)).await;
 
         assert!(limits.is_limited());
-        assert!(envelope.envelope().is_empty()); // obviously
+        assert!(envelope.is_empty()); // obviously
         mock.lock().await.assert_call(DataCategory::Transaction, 1);
     }
 
     #[tokio::test]
     #[cfg(feature = "processing")]
     async fn test_enforce_limit_assumed_attachments() {
-        let mut envelope = envelope![Attachment, Attachment];
+        let envelope = envelope![Event, Attachment, Attachment];
 
         let mock = mock_limiter(&[DataCategory::Error]);
-        let (_, limits) =
-            enforce_and_apply(mock.clone(), &mut envelope, Some(DataCategory::Error)).await;
+        let (envelope, _, limits) =
+            enforce_and_apply(mock.clone(), envelope, Some(DataCategory::Error)).await;
 
         assert!(limits.is_limited());
-        assert!(envelope.envelope().is_empty());
+        assert!(envelope.is_empty());
         mock.lock().await.assert_call(DataCategory::Error, 1);
     }
 
     #[tokio::test]
     async fn test_enforce_transaction() {
-        let mut envelope = envelope![Transaction];
+        let envelope = envelope![Transaction];
 
         let mock = mock_limiter(&[DataCategory::Transaction]);
-        let (enforcement, limits) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
+        let (_, enforcement, limits) = enforce_and_apply(mock.clone(), envelope, None).await;
 
         assert!(limits.is_limited());
         assert!(enforcement.event_indexed.is_active());
@@ -2043,8 +2008,12 @@ mod tests {
 
     #[tokio::test]
     async fn test_enforce_transaction_non_indexed() {
-        let mut envelope = envelope![Transaction, Profile];
-        let scoping = envelope.scoping();
+        let envelope = envelope![Transaction, Profile];
+        let scoping = envelope
+            .headers()
+            .meta()
+            .get_partial_scoping()
+            .into_scoping();
 
         let mock = mock_limiter(&[DataCategory::TransactionIndexed]);
 
@@ -2056,11 +2025,7 @@ mod tests {
                 mock.check(s, q)
             }
         });
-        let (enforcement, limits) = limiter
-            .compute(envelope.envelope_mut(), &scoping)
-            .await
-            .unwrap();
-        enforcement.clone().apply_with_outcomes(&mut envelope);
+        let (enforcement, limits) = limiter.compute(&envelope, &scoping).await.unwrap();
 
         assert!(!limits.is_limited());
         assert!(!enforcement.event_indexed.is_active());
@@ -2076,26 +2041,24 @@ mod tests {
 
     #[tokio::test]
     async fn test_enforce_transaction_no_indexing_quota() {
-        let mut envelope = envelope![Transaction];
+        let envelope = envelope![Transaction];
 
         let mock = mock_limiter(&[DataCategory::TransactionIndexed]);
-        let (enforcement, limits) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
+        let (_, enforcement, limits) = enforce_and_apply(mock.clone(), envelope, None).await;
 
-        assert!(limits.is_limited());
-        assert!(enforcement.event_indexed.is_active());
+        assert!(!limits.is_limited());
+        assert!(!enforcement.event_indexed.is_active());
         assert!(!enforcement.event.is_active());
         mock.lock().await.assert_call(DataCategory::Transaction, 1);
-        mock.lock()
-            .await
-            .assert_call(DataCategory::TransactionIndexed, 1);
+        mock.lock().await.assert_call(DataCategory::Span, 1);
     }
 
     #[tokio::test]
     async fn test_enforce_transaction_attachment_enforced() {
-        let mut envelope = envelope![Transaction, Attachment];
+        let envelope = envelope![Transaction, Attachment];
 
         let mock = mock_limiter(&[DataCategory::Transaction]);
-        let (enforcement, _) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
+        let (_, enforcement, _) = enforce_and_apply(mock.clone(), envelope, None).await;
 
         assert!(enforcement.event.is_active());
         assert!(enforcement.attachments_limits.event.is_active());
@@ -2111,10 +2074,10 @@ mod tests {
 
     #[tokio::test]
     async fn test_enforce_transaction_profile_enforced() {
-        let mut envelope = envelope![Transaction, Profile];
+        let envelope = envelope![Transaction, Profile];
 
         let mock = mock_limiter(&[DataCategory::Transaction]);
-        let (enforcement, _) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
+        let (_, enforcement, _) = enforce_and_apply(mock.clone(), envelope, None).await;
 
         assert!(enforcement.event.is_active());
         assert!(enforcement.profiles.is_active());
@@ -2136,10 +2099,10 @@ mod tests {
     #[tokio::test]
     async fn test_enforce_transaction_standalone_profile_enforced() {
         // When the transaction is sampled, the profile survives as standalone.
-        let mut envelope = envelope![Profile];
+        let envelope = envelope![Profile];
 
         let mock = mock_limiter(&[DataCategory::Transaction]);
-        let (enforcement, _) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
+        let (_, enforcement, _) = enforce_and_apply(mock.clone(), envelope, None).await;
 
         assert!(enforcement.profiles.is_active());
         mock.lock().await.assert_call(DataCategory::Profile, 1);
@@ -2156,37 +2119,30 @@ mod tests {
 
     #[tokio::test]
     async fn test_enforce_transaction_attachment_enforced_indexing_quota() {
-        let mut envelope = envelope![Transaction, Attachment];
-        set_extracted(envelope.envelope_mut(), ItemType::Transaction);
+        let envelope = envelope![Transaction, Attachment];
 
         let mock = mock_limiter(&[DataCategory::TransactionIndexed]);
-        let (enforcement, _) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
+        let (_, enforcement, _) = enforce_and_apply(mock.clone(), envelope, None).await;
 
         assert!(!enforcement.event.is_active());
-        assert!(enforcement.event_indexed.is_active());
-        assert!(enforcement.attachments_limits.event.is_active());
+        assert!(!enforcement.event_indexed.is_active());
+        assert!(!enforcement.attachments_limits.event.is_active());
         mock.lock().await.assert_call(DataCategory::Transaction, 1);
+        mock.lock().await.assert_call(DataCategory::Span, 1);
+        mock.lock().await.assert_call(DataCategory::Attachment, 10);
         mock.lock()
             .await
-            .assert_call(DataCategory::TransactionIndexed, 1);
+            .assert_call(DataCategory::AttachmentItem, 1);
 
-        assert_eq!(
-            get_outcomes(enforcement),
-            vec![
-                (DataCategory::TransactionIndexed, 1),
-                (DataCategory::Attachment, 10),
-                (DataCategory::AttachmentItem, 1),
-                (DataCategory::SpanIndexed, 1),
-            ]
-        );
+        assert_eq!(get_outcomes(enforcement), vec![]);
     }
 
     #[tokio::test]
     async fn test_enforce_span() {
-        let mut envelope = envelope![Span, Span];
+        let envelope = envelope![Span, Span];
 
         let mock = mock_limiter(&[DataCategory::Span]);
-        let (enforcement, limits) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
+        let (_, enforcement, limits) = enforce_and_apply(mock.clone(), envelope, None).await;
 
         assert!(limits.is_limited());
         assert!(enforcement.spans_indexed.is_active());
@@ -2201,41 +2157,17 @@ mod tests {
 
     #[tokio::test]
     async fn test_enforce_span_no_indexing_quota() {
-        let mut envelope = envelope![Span, Span];
+        let envelope = envelope![Span, Span];
 
         let mock = mock_limiter(&[DataCategory::SpanIndexed]);
-        let (enforcement, limits) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
+        let (_, enforcement, limits) = enforce_and_apply(mock.clone(), envelope, None).await;
 
-        assert!(limits.is_limited());
-        assert!(enforcement.spans_indexed.is_active());
+        assert!(!limits.is_limited());
+        assert!(!enforcement.spans_indexed.is_active());
         assert!(!enforcement.spans.is_active());
         mock.lock().await.assert_call(DataCategory::Span, 2);
-        mock.lock().await.assert_call(DataCategory::SpanIndexed, 2);
 
-        assert_eq!(
-            get_outcomes(enforcement),
-            vec![(DataCategory::SpanIndexed, 2)]
-        );
-    }
-
-    #[tokio::test]
-    async fn test_enforce_span_metrics_extracted_no_indexing_quota() {
-        let mut envelope = envelope![Span, Span];
-        set_extracted(envelope.envelope_mut(), ItemType::Span);
-
-        let mock = mock_limiter(&[DataCategory::SpanIndexed]);
-        let (enforcement, limits) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
-
-        assert!(limits.is_limited());
-        assert!(enforcement.spans_indexed.is_active());
-        assert!(!enforcement.spans.is_active());
-        mock.lock().await.assert_call(DataCategory::Span, 2);
-        mock.lock().await.assert_call(DataCategory::SpanIndexed, 2);
-
-        assert_eq!(
-            get_outcomes(enforcement),
-            vec![(DataCategory::SpanIndexed, 2)]
-        );
+        assert_eq!(get_outcomes(enforcement), vec![]);
     }
 
     #[test]
@@ -2270,13 +2202,13 @@ mod tests {
 
     #[tokio::test]
     async fn test_enforce_limit_logs_count() {
-        let mut envelope = envelope![Log, Log];
+        let envelope = envelope![Log, Log];
 
         let mock = mock_limiter(&[DataCategory::LogItem]);
-        let (enforcement, limits) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
+        let (envelope, enforcement, limits) = enforce_and_apply(mock.clone(), envelope, None).await;
 
         assert!(limits.is_limited());
-        assert_eq!(envelope.envelope().len(), 0);
+        assert_eq!(envelope.len(), 0);
         mock.lock().await.assert_call(DataCategory::LogItem, 2);
 
         assert_eq!(
@@ -2287,13 +2219,13 @@ mod tests {
 
     #[tokio::test]
     async fn test_enforce_limit_logs_bytes() {
-        let mut envelope = envelope![Log, Log];
+        let envelope = envelope![Log, Log];
 
         let mock = mock_limiter(&[DataCategory::LogByte]);
-        let (enforcement, limits) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
+        let (envelope, enforcement, limits) = enforce_and_apply(mock.clone(), envelope, None).await;
 
         assert!(limits.is_limited());
-        assert_eq!(envelope.envelope().len(), 0);
+        assert_eq!(envelope.len(), 0);
         mock.lock().await.assert_call(DataCategory::LogItem, 2);
         mock.lock().await.assert_call(DataCategory::LogByte, 20);
 
@@ -2319,22 +2251,19 @@ mod tests {
             RateLimitTestCase {
                 name: "span_indexed_limit",
                 denied_categories: &[DataCategory::SpanIndexed],
-                expect_attachment_limit_active: true,
-                expected_limiter_calls: &[(DataCategory::Span, 0), (DataCategory::SpanIndexed, 0)],
-                expected_outcomes: &[
+                expect_attachment_limit_active: false,
+                expected_limiter_calls: &[
+                    (DataCategory::Span, 0),
                     (DataCategory::Attachment, 7),
                     (DataCategory::AttachmentItem, 1),
                 ],
+                expected_outcomes: &[],
             },
             RateLimitTestCase {
                 name: "attachment_limit",
                 denied_categories: &[DataCategory::Attachment],
                 expect_attachment_limit_active: true,
-                expected_limiter_calls: &[
-                    (DataCategory::Span, 0),
-                    (DataCategory::SpanIndexed, 0),
-                    (DataCategory::Attachment, 7),
-                ],
+                expected_limiter_calls: &[(DataCategory::Span, 0), (DataCategory::Attachment, 7)],
                 expected_outcomes: &[
                     (DataCategory::Attachment, 7),
                     (DataCategory::AttachmentItem, 1),
@@ -2346,7 +2275,6 @@ mod tests {
                 expect_attachment_limit_active: true,
                 expected_limiter_calls: &[
                     (DataCategory::Span, 0),
-                    (DataCategory::SpanIndexed, 0),
                     (DataCategory::Attachment, 7),
                     (DataCategory::AttachmentItem, 1),
                 ],
@@ -2361,7 +2289,6 @@ mod tests {
                 expect_attachment_limit_active: false,
                 expected_limiter_calls: &[
                     (DataCategory::Span, 0),
-                    (DataCategory::SpanIndexed, 0),
                     (DataCategory::Attachment, 7),
                     (DataCategory::AttachmentItem, 1),
                 ],
@@ -2373,7 +2300,6 @@ mod tests {
                 expect_attachment_limit_active: false,
                 expected_limiter_calls: &[
                     (DataCategory::Span, 0),
-                    (DataCategory::SpanIndexed, 0),
                     (DataCategory::Attachment, 7),
                     (DataCategory::AttachmentItem, 1),
                 ],
@@ -2385,7 +2311,6 @@ mod tests {
                 expect_attachment_limit_active: false,
                 expected_limiter_calls: &[
                     (DataCategory::Span, 0),
-                    (DataCategory::SpanIndexed, 0),
                     (DataCategory::Attachment, 7),
                     (DataCategory::AttachmentItem, 1),
                 ],
@@ -2402,12 +2327,10 @@ mod tests {
         } in test_cases
         {
             let mut envelope = envelope![];
-            envelope
-                .envelope_mut()
-                .add_item(trace_attachment_item(7, Some(ParentId::SpanId(None))));
+            envelope.add_item(trace_attachment_item(7, Some(ParentId::SpanId(None))));
 
             let mock = mock_limiter(denied_categories);
-            let (enforcement, _) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
+            let (_, enforcement, _) = enforce_and_apply(mock.clone(), envelope, None).await;
 
             for &(category, quantity) in *expected_limiter_calls {
                 mock.lock().await.assert_call(category, quantity);
@@ -2451,22 +2374,17 @@ mod tests {
                 name: "span_indexed_limit",
                 denied_categories: &[DataCategory::SpanIndexed, DataCategory::Attachment],
                 expect_attachment_limit_active: true,
-                expected_limiter_calls: &[(DataCategory::Span, 1), (DataCategory::SpanIndexed, 1)],
+                expected_limiter_calls: &[(DataCategory::Span, 1), (DataCategory::Attachment, 7)],
                 expected_outcomes: &[
                     (DataCategory::Attachment, 7),
                     (DataCategory::AttachmentItem, 1),
-                    (DataCategory::SpanIndexed, 1),
                 ],
             },
             RateLimitTestCase {
                 name: "attachment_limit",
                 denied_categories: &[DataCategory::Attachment],
                 expect_attachment_limit_active: true,
-                expected_limiter_calls: &[
-                    (DataCategory::Span, 1),
-                    (DataCategory::SpanIndexed, 1),
-                    (DataCategory::Attachment, 7),
-                ],
+                expected_limiter_calls: &[(DataCategory::Span, 1), (DataCategory::Attachment, 7)],
                 expected_outcomes: &[
                     (DataCategory::Attachment, 7),
                     (DataCategory::AttachmentItem, 1),
@@ -2478,7 +2396,6 @@ mod tests {
                 expect_attachment_limit_active: true,
                 expected_limiter_calls: &[
                     (DataCategory::Span, 1),
-                    (DataCategory::SpanIndexed, 1),
                     (DataCategory::Attachment, 7),
                     (DataCategory::AttachmentItem, 1),
                 ],
@@ -2493,7 +2410,6 @@ mod tests {
                 expect_attachment_limit_active: false,
                 expected_limiter_calls: &[
                     (DataCategory::Span, 1),
-                    (DataCategory::SpanIndexed, 1),
                     (DataCategory::Attachment, 7),
                     (DataCategory::AttachmentItem, 1),
                 ],
@@ -2505,7 +2421,6 @@ mod tests {
                 expect_attachment_limit_active: false,
                 expected_limiter_calls: &[
                     (DataCategory::Span, 1),
-                    (DataCategory::SpanIndexed, 1),
                     (DataCategory::Attachment, 7),
                     (DataCategory::AttachmentItem, 1),
                 ],
@@ -2517,7 +2432,6 @@ mod tests {
                 expect_attachment_limit_active: false,
                 expected_limiter_calls: &[
                     (DataCategory::Span, 1),
-                    (DataCategory::SpanIndexed, 1),
                     (DataCategory::Attachment, 7),
                     (DataCategory::AttachmentItem, 1),
                 ],
@@ -2534,12 +2448,10 @@ mod tests {
         } in test_cases
         {
             let mut envelope = envelope![Span];
-            envelope
-                .envelope_mut()
-                .add_item(trace_attachment_item(7, Some(ParentId::SpanId(None))));
+            envelope.add_item(trace_attachment_item(7, Some(ParentId::SpanId(None))));
 
             let mock = mock_limiter(denied_categories);
-            let (enforcement, _) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
+            let (_, enforcement, _) = enforce_and_apply(mock.clone(), envelope, None).await;
 
             for &(category, quantity) in *expected_limiter_calls {
                 mock.lock().await.assert_call(category, quantity);
@@ -2568,22 +2480,6 @@ mod tests {
     async fn test_enforce_transaction_span_attachment() {
         let test_cases = &[
             RateLimitTestCase {
-                name: "span_limit",
-                denied_categories: &[DataCategory::Span],
-                expect_attachment_limit_active: true,
-                expected_limiter_calls: &[
-                    (DataCategory::Transaction, 1),
-                    (DataCategory::TransactionIndexed, 1),
-                    (DataCategory::Span, 1),
-                ],
-                expected_outcomes: &[
-                    (DataCategory::Attachment, 7),
-                    (DataCategory::AttachmentItem, 1),
-                    (DataCategory::Span, 1),
-                    (DataCategory::SpanIndexed, 1),
-                ],
-            },
-            RateLimitTestCase {
                 name: "transaction_limit",
                 denied_categories: &[DataCategory::Transaction],
                 expect_attachment_limit_active: true,
@@ -2603,9 +2499,7 @@ mod tests {
                 expect_attachment_limit_active: false,
                 expected_limiter_calls: &[
                     (DataCategory::Transaction, 1),
-                    (DataCategory::TransactionIndexed, 1),
                     (DataCategory::Span, 1),
-                    (DataCategory::SpanIndexed, 1),
                     (DataCategory::Attachment, 7),
                     (DataCategory::AttachmentItem, 1),
                 ],
@@ -2617,9 +2511,7 @@ mod tests {
                 expect_attachment_limit_active: false,
                 expected_limiter_calls: &[
                     (DataCategory::Transaction, 1),
-                    (DataCategory::TransactionIndexed, 1),
                     (DataCategory::Span, 1),
-                    (DataCategory::SpanIndexed, 1),
                     (DataCategory::Attachment, 7),
                     (DataCategory::AttachmentItem, 1),
                 ],
@@ -2636,12 +2528,10 @@ mod tests {
         } in test_cases
         {
             let mut envelope = envelope![Transaction];
-            envelope
-                .envelope_mut()
-                .add_item(trace_attachment_item(7, Some(ParentId::SpanId(None))));
+            envelope.add_item(trace_attachment_item(7, Some(ParentId::SpanId(None))));
 
             let mock = mock_limiter(denied_categories);
-            let (enforcement, _) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
+            let (_, enforcement, _) = enforce_and_apply(mock.clone(), envelope, None).await;
 
             for &(category, quantity) in *expected_limiter_calls {
                 mock.lock().await.assert_call(category, quantity);
@@ -2723,12 +2613,10 @@ mod tests {
         } in test_cases
         {
             let mut envelope = envelope![];
-            envelope
-                .envelope_mut()
-                .add_item(trace_attachment_item(7, None));
+            envelope.add_item(trace_attachment_item(7, None));
 
             let mock = mock_limiter(denied_categories);
-            let (enforcement, _) = enforce_and_apply(mock.clone(), &mut envelope, None).await;
+            let (_, enforcement, _) = enforce_and_apply(mock.clone(), envelope, None).await;
 
             for &(category, quantity) in *expected_limiter_calls {
                 mock.lock().await.assert_call(category, quantity);